RORDEV-1374 move mock to separate file and use stub
