+++ e2e-tests/cypress/support/commands.ts
@@ -1,4 +1,5 @@
 import '@testing-library/cypress/add-commands';
+import { ResizeObserverMock } from './mocks/ResizeObserverMock';
 
 Cypress.Commands.add('kbnPost', ({ endpoint, credentials, payload, currentGroupHeader }, ...args) => {
   cy.kbnRequest({
@@ -150,45 +151,5 @@ Cypress.on('uncaught:exception', (err, runnable) => {
 });
 
 Cypress.on('window:before:load', win => {
-  class ResizeObserverMock {
-    private callback: ResizeObserverCallback;
-    private elements: Set<Element>;
-
-    constructor(callback: ResizeObserverCallback) {
-      // Debounce the callback to prevent ResizeObserver loops
-      this.callback = Cypress._.debounce(callback, 200);
-      this.elements = new Set();
-    }
-
-    observe(element: Element) {
-      this.elements.add(element);
-
-      // Simulate a resize event with a valid ResizeObserverEntry structure
-      setTimeout(() => {
-        if (this.elements.has(element)) {
-          const entries: ResizeObserverEntry[] = [
-            {
-              target: element,
-              contentRect: element.getBoundingClientRect(),
-              borderBoxSize: [{ inlineSize: element.clientWidth, blockSize: element.clientHeight }],
-              contentBoxSize: [{ inlineSize: element.clientWidth, blockSize: element.clientHeight }],
-              devicePixelContentBoxSize: [{ inlineSize: element.clientWidth, blockSize: element.clientHeight }]
-            } as ResizeObserverEntry
-          ];
-          this.callback(entries, this);
-        }
-      }, 200);
-    }
-
-    unobserve(element: Element) {
-      this.elements.delete(element);
-    }
-
-    disconnect() {
-      this.elements.clear();
-    }
-  }
-
-  // Override ResizeObserver in the test environment
-  win.ResizeObserver = ResizeObserverMock as unknown as typeof ResizeObserver;
+  cy.stub(win, 'ResizeObserver').returns(ResizeObserverMock);
 });
